Hide lockscreen items [1/2]
This adds the ability to add items such as data, clock and alarm in the
lockscreen.

Thanks to beanstown106 for the tip on the alarm text mod

Change-Id: I42c7ad0c3de0b41a653b9b687acfe9536bf80a44

Always keep clock/date & alarm visible while dozing

When we introduced this mod in 8.1, we didn't account for ambient display

DirtyUnicorns/android_frameworks_base@deb4cf4

Since we didn't account for AD, if you turned off these items you wouldn't see them on
the lockscreen while the screen is on but you also wouldn't see them while dozing.

This was just bad design. The user should ALWAYS see these items while dozing. If they don't
then the user might get the impression that ambient display doesn't work. Believe me, it happened
to me because I forgot I had those items toggled off. All you see is a black screen if there's no
notifications.

We now check for when the device is dozing/screen off and the switch respects dozing.

Steps to reproduce the fail:
- Turn on Ambient display
- Go into Settings/Dirty Tweaks/Lockscreen items and turn off any or all the items
- Enter doze mode

Thanks to Ezio for his method to check Doze. (I modified it to check for screen off)

Change-Id: Ic9e62e7f7399ed321b923de7ddb939c76f6caa2d